Refactor: move kafka stuff to sub-project.
