+++ dependencies.gradle
@@ -6,10 +6,11 @@ ext {
   libs = [
     TEST : ['junit:junit:4.12', 'org.mockito:mockito-core:1.10.19'],
     SPARK : [
-        "org.apache.spark:spark-core_$SPARK_SCALA_VERSION:$SPARK_VERSION",
         "org.apache.spark:spark-streaming_$SPARK_SCALA_VERSION:$SPARK_VERSION",
-        "org.apache.spark:spark-streaming-kafka-0-8_$SPARK_SCALA_VERSION:$SPARK_VERSION"
-      ],
+        "org.apache.spark:spark-core_$SPARK_SCALA_VERSION:$SPARK_VERSION"
+    ],
+    KAFKA_0_8 : "org.apache.spark:spark-streaming-kafka-0-8_$SPARK_SCALA_VERSION:$SPARK_VERSION",
+    KAFKA_0_10 : "org.apache.spark:spark-streaming-kafka-0-10_$SPARK_SCALA_VERSION:$SPARK_VERSION",
     COMMONS_CLI :'commons-cli:commons-cli:1.2',
     COMMONS_LANG : 'org.apache.commons:commons-lang3:3.3.2',
     AWS : ['org.apache.hadoop:hadoop-aws:2.7.1', 'com.amazonaws:aws-java-sdk:1.7.4'],
