+++ di-providers/kafka-0-8/build.gradle
@@ -1,0 +1,17 @@
+description = 'Kafka 0.8 provider for Spark Streamming'
+
+ext {
+  projectName = 'nightfall-di-kafka-0-8'
+}
+
+jar {
+  appendix = 'di-kafka-0_8'
+}
+
+dependencies {
+  compile  project(':di')
+  provided libs.SPARK, libs.COMMONS_CLI, libs.AWS, libs.SLF4J, libs.KAFKA_0_8
+          // Provided due lack of support for optional dependencies from gradle: https://issues.gradle.org/browse/GRADLE-1749
+          project(':persistence/cassandra')
+  testCompile libs.TEST
+}
