+++ di-providers/kafka-0-8/src/main/java/com/elo7/nightfall/di/providers/kafka/KafkaSimple.java
