+++ di-providers/kafka-0-8/src/main/java/com/elo7/nightfall/di/providers/kafka/TopicNotFoundException.java
@@ -1,0 +1,9 @@
+package com.elo7.nightfall.di.providers.kafka;
+
+public class TopicNotFoundException extends RuntimeException {
+	private static final long serialVersionUID = 1L;
+
+	public TopicNotFoundException(String message) {
+		super(message);
+	}
+}
