+++ di-providers/kafka-0-8/src/main/java/com/elo7/nightfall/di/providers/kafka/simple/KafkaAutoOffsetReset.java
@@ -1,6 +1,6 @@
 package com.elo7.nightfall.di.providers.kafka.simple;
 
 public enum KafkaAutoOffsetReset {
-    LARGEST,
-    SMALLEST,
+	LARGEST,
+	SMALLEST,
 }
