+++ di/src/test/resources/nightfall.properties
@@ -1,35 +1,50 @@
 # Batch
-batch.history.enabled=false
-
-# Monitoring config
-reporter.enabled=false
+batch.history.enabled=false;
 
 # Batch - Job History
-batch.cassandra.host=localhost
+batch.cassandra.host=cassandra
 batch.cassandra.port=9042
 batch.cassandra.user=
 batch.cassandra.password=
 batch.cassandra.keyspace=nightfall
 batch.cassandra.datacenter=
-batch.history.ttl.days=7
+batch.history.ttl.days=7;
+
+# S3RDD
+s3.access.key=
+s3.secret.key=
+s3.path=meu/caminho/wild*.gz.
+s3.bucket=nightfall.elo7.com.
+s3.protocol=s3a://
 
 # Kafka Simple
-kafka.brokers=localhost:9092
-kafka.topics=events
+kafka.brokers=192.168.99.100:9092
+kafka.topics=marketplace_events
 kafka.offset.persistent=false
 kafka.auto.offset.reset=smallest
+kafka.simple.repository.class=com.elo7.nightfall.di.providers.kafka.topics.CassandraKafkaTopicRepository
 
 # Kafka Simple - Offset History
-kafka.cassandra.hosts=localhost
+kafka.cassandra.hosts=cassandra
 kafka.cassandra.keyspace=kafka
 
 # Kafka High Level
 kafka.group=NightfallGroup
-kafka.topics.map=events
-kafka.zookeeper=localhost:2181/kafka
+kafka.topics.map=marketplace_events
+kafka.zookeeper=zookeeper:2181/kafka
 kafka.default.topic.partitions=1
 
 # Streaming
-stream.graceful.shutdown=true
-stream.batch.interval.ms=1000
+stream.graceful.shutdown=true.
+stream.batch.interval.ms=1000.
 stream.writeAheadLog.enable=false
+stream.checkpoint.directory=
+stream.back.pressure.enabled=true
+stream.max.rate.per.second =1000
+
+# Monitoring config
+reporter.statsd.host=localhost
+reporter.statsd.port=8125
+reporter.statsd.prefix=dev.spark
+reporter.enabled=false
+reporter.class=com.elo7.nightfall.di.providers.reporter.statsd.StatsDReporterFactory
